Rename types/server.d.ts to tyeps/index.d.ts
Update imports as needed
